+++ dune/localfunctions/lagrange.hh
@@ -3,12 +3,38 @@
 #ifndef DUNE_LOCALFUNCTIONS_LAGRANGE_HH
 #define DUNE_LOCALFUNCTIONS_LAGRANGE_HH
 
+/** \file
+ * \brief Convenience header that includes all implementations of Lagrange finite elements
+ */
+
+// Headers for Lagrange elements with run-time order
 #include <dune/localfunctions/utility/localfiniteelement.hh>
 #include <dune/localfunctions/utility/dglocalcoefficients.hh>
 
 #include <dune/localfunctions/lagrange/lagrangecoefficients.hh>
 #include <dune/localfunctions/lagrange/interpolation.hh>
 #include <dune/localfunctions/lagrange/lagrangebasis.hh>
+
+// Headers for Lagrange elements with compile-time order
+#include <dune/localfunctions/lagrange/p0.hh>
+#include <dune/localfunctions/lagrange/p1.hh>
+#include <dune/localfunctions/lagrange/p2.hh>
+#include <dune/localfunctions/lagrange/p23d.hh>
+#include <dune/localfunctions/lagrange/pk.hh>
+#include <dune/localfunctions/lagrange/pk1d.hh>
+#include <dune/localfunctions/lagrange/pk2d.hh>
+#include <dune/localfunctions/lagrange/pk3d.hh>
+#include <dune/localfunctions/lagrange/pq22d.hh>
+
+#include <dune/localfunctions/lagrange/q1.hh>
+#include <dune/localfunctions/lagrange/q2.hh>
+#include <dune/localfunctions/lagrange/qk.hh>
+
+#include <dune/localfunctions/lagrange/prismp1.hh>
+#include <dune/localfunctions/lagrange/prismp2.hh>
+#include <dune/localfunctions/lagrange/pyramidp1.hh>
+#include <dune/localfunctions/lagrange/pyramidp2.hh>
+
 
 namespace Dune
 {